src/requirements.txt
pandas 
SentenceTransformers
pinecone-client